+++ Source/UI/Activation/Process/ActivationUIProcess.swift
@@ -27,7 +27,6 @@ public protocol ActivationUIProvider: class {
     func instantiateNavigationController(with rootController: UIViewController) -> UINavigationController?
     
     var uiDataProvider: ActivationUIDataProvider { get }
-    var recoveryUIProvider: RecoveryUIProvider { get }
     var authenticationUIProvider: AuthenticationUIProvider { get }
     var actionFeedback: LimeAuthActionFeedback? { get }
 }
