+++ Source/UIResources/Authentication/DefaultAuthenticationResourcesProvider.swift
@@ -23,14 +23,20 @@ internal class DefaultAuthenticationResourcesProvider: AuthenticationUIProvider,
     public let localization: GenericLocalizationProvider
     public lazy var actionFeedback: LimeAuthActionFeedback? = LimeAuthActionFeedback()
     
-    public init(bundle: Bundle? = nil, localizationProvider: GenericLocalizationProvider?) {
+    public init(bundle: Bundle? = nil, localizationProvider: GenericLocalizationProvider?, recoveryClosure: @escaping @autoclosure ()->RecoveryUIProvider) {
         self.bundle = bundle ?? .limeAuthResourcesBundle
         self.localization = localizationProvider ?? SystemLocalizationProvider(tableName: "LimeAuth", bundle: .limeAuthResourcesBundle)
+        self.recoveryUIProviderClosure = recoveryClosure
     }
     
     private var storyboard: UIStoryboard {
         return UIStoryboard(name: "Authentication", bundle: bundle)
     }
+    
+    private let recoveryUIProviderClosure: ()->RecoveryUIProvider
+    public lazy var recoveryUIProvider: RecoveryUIProvider = {
+        return recoveryUIProviderClosure()
+    }()
     
     //
